Add files via upload
Adding code to filter data by target years and zip codes, then to take that into a zip code-indexed df where the calls by year are the values. then extracted that into the "Training.csv". Also adding the schema I used to assign out the "Department" column as a way of categorizing call type, still working to make that functional.